[FIX] 소비등록 반복 결과 버튼 수정
