+++ Money-Planner/Money-Planner/View/Consume/modal/ChooseDateView.swift
@@ -1,34 +1,30 @@
 //
-//  ChooseDayView.swift
+//  ChooseDateView.swift
 //  Money-Planner
 //
-//  Created by p_kxn_g on 1/27/24.
+//  Created by p_kxn_g on 1/28/24.
 //
 
+import Foundation
 import UIKit
 
 // Protocol to handle interactions or events related to ChooseDayView
-protocol ChooseDayViewDelegate: AnyObject {
-    func chooseIntervalDay(_ view: UIViewController)
-    func addEndDay(_ view : UIStackView)
-    func removeEndDay()
+protocol ChooseDateViewDelegate: AnyObject {
+    func chooseIntervalDate(_ view: UIViewController)
+    func addEndDate(_ view : UIStackView)
+    func removeEndDate()
     // Define methods or properties as needed
 }
 
-class ChooseDayView: UIView {
-    
+class ChooseDateView: UIView {
     // Weak reference to the delegate
-    weak var delegate: ChooseDayViewDelegate?
+    weak var delegate: ChooseDateViewDelegate?
+    var weekIBtnConstraint  : NSLayoutConstraint?
+
     var repeatEndChecked = true
-    let currentDate = Date()
-    let dateFormatter = DateFormatter()
-    let button1 = OneDayButton(title: "월", buttonBool: true)
-    let button2 = OneDayButton(title: "화", buttonBool: true)
-    let button3 = OneDayButton(title: "수", buttonBool: true)
-    let button4 = OneDayButton(title: "목", buttonBool: true)
-    let button5 = OneDayButton(title: "금", buttonBool: true)
-    let button6 = OneDayButton(title: "토", buttonBool: true)
-    let button7 = OneDayButton(title: "일", buttonBool: true)
+    private let currentDate = Date()
+    private let dateFormatter = DateFormatter()
+  
     lazy var dateString: String = {
         dateFormatter.dateFormat = "yyyy년 MM월 dd일"
         // 일년 뒤 날짜.
@@ -39,18 +35,16 @@ class ChooseDayView: UIView {
         }
         
     }()
+    lazy var dateString2: String = {
+        dateFormatter.dateFormat = "dd일"
+        return dateFormatter.string(from: currentDate)
+        
+    }()
+    
     // Other properties and methods of ChooseDayView
     private let backButton = UIButton()
     private let titleLabel = MPLabel()
-    private let weekButtons: UIStackView = {
-            let stackView = UIStackView()
-            stackView.axis = .horizontal
-            stackView.distribution = .fillEqually
-        stackView.backgroundColor = .red
-            stackView.spacing = 8
-            return stackView
-        }()
-    
+
     private let weekInterval : UIStackView = {
             let stackView = UIStackView()
         // >> stackView.backgroundColor = .blue
@@ -58,17 +52,22 @@ class ChooseDayView: UIView {
             stackView.spacing = 8
             return stackView
         }()
-    let weekIntervalButton = OneDayButton(title: "1",buttonBool: false)
-
+  
+    lazy var weekIntervalButton: OneDayButton = {
+            // Now you can use dateString2 here
+            let button = OneDayButton(title: dateString2, buttonBool: false)
+            button.titleLabel?.font = .mpFont16R()
+            return button
+        }()
     private let repeatEndDate : UIStackView = {
             let stackView = UIStackView()
-        // >> stackView.backgroundColor = .cyan
+        //>> stackView.backgroundColor = .cyan
             stackView.axis = .horizontal
             stackView.spacing = 8
         stackView.alignment = .leading
             return stackView
         }()
-    private let repeatEndDate2 : UIStackView = {
+    let repeatEndDate2 : UIStackView = {
             let stackView = UIStackView()
         // >> stackView.backgroundColor = .mpWhite
             stackView.axis = .horizontal
@@ -77,7 +76,7 @@ class ChooseDayView: UIView {
             return stackView
         }()
     
-    let RepeatEndDate2Label : UILabel = {
+     let RepeatEndDate2Label : UILabel = {
         let label = UILabel()
         label.text = "까지"
         label.font = UIFont.mpFont16R()
@@ -101,90 +100,85 @@ class ChooseDayView: UIView {
         button.setTitleColor(.mpDarkGray, for: .normal)
         button.translatesAutoresizingMaskIntoConstraints = false
         button.isUserInteractionEnabled = true  //클릭 활성화
+       
         return button
         
     }()
     
     override init(frame: CGRect  = .zero) {
+        
+        //let view1 = UIView(frame: CGRect(x: 0, y: 0, width: 313, height: 171))
         super.init(frame: frame)
-        //backgroundColor = .blue
+        
         setupContainer()
-        setupWeekButtons()
         setupWeekInterval()
         setupRepeatEndDate()
-        //setupRepeatEndDateButton()
+        //
+        //backgroundColor = .blue
+        //ㄹweekInterval.backgroundColor = .green
+
     }
     
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
     private func setupContainer() {
+        
        
-        weekButtons.translatesAutoresizingMaskIntoConstraints = false
         weekInterval.translatesAutoresizingMaskIntoConstraints = false
         repeatEndDate.translatesAutoresizingMaskIntoConstraints = false
-        weekIntervalButton.titleLabel?.font = .mpFont16R()
-        addSubview(weekButtons)
         addSubview(weekInterval)
         addSubview(repeatEndDate)
         //repeatEndDate.backgroundColor = .yellow
            NSLayoutConstraint.activate([
             // 높이 설정
-            weekButtons.heightAnchor.constraint(equalToConstant: 38),
+            
             weekInterval.heightAnchor.constraint(equalToConstant: 38),
             repeatEndDate.heightAnchor.constraint(equalToConstant: 24),
                //weekButtons.bottomAnchor.constraint(equalTo: weekInterval.topAnchor),
-               weekButtons.topAnchor.constraint(equalTo: topAnchor),
-            weekButtons.leadingAnchor.constraint(equalTo: leadingAnchor),
-            weekButtons.trailingAnchor.constraint(equalTo: trailingAnchor),
-            weekButtons.centerXAnchor.constraint(equalTo: centerXAnchor),
+              
             
-               weekInterval.topAnchor.constraint(equalTo: weekButtons.bottomAnchor,constant: 16),
+               weekInterval.topAnchor.constraint(equalTo: topAnchor),
                weekInterval.leadingAnchor.constraint(equalTo: leadingAnchor),
                weekInterval.trailingAnchor.constraint(equalTo: trailingAnchor),
                //repeatEndDate.bottomAnchor.constraint(equalTo: bottomAnchor),
                repeatEndDate.topAnchor.constraint(equalTo: weekInterval.bottomAnchor,constant: 16),
                repeatEndDate.leadingAnchor.constraint(equalTo: leadingAnchor),
                repeatEndDate.trailingAnchor.constraint(equalTo: trailingAnchor),
-            weekIntervalButton.widthAnchor.constraint(equalToConstant: 38)
            ])
     }
     
-    private func setupWeekButtons(){
-
-        NSLayoutConstraint.activate([
-//            button1.widthAnchor.constraint(equalToConstant: 38),
-//            button2.widthAnchor.constraint(equalToConstant: 38),
-//            button3.widthAnchor.constraint(equalToConstant: 38),
-//            button4.widthAnchor.constraint(equalToConstant: 38),
-//            button5.widthAnchor.constraint(equalToConstant: 38),
-//            button6.widthAnchor.constraint(equalToConstant: 38),
-//            button7.widthAnchor.constraint(equalToConstant: 38),
-        ])
-        weekButtons.addArrangedSubview(button1)
-        weekButtons.addArrangedSubview(button2)
-        weekButtons.addArrangedSubview(button3)
-        weekButtons.addArrangedSubview(button4)
-        weekButtons.addArrangedSubview(button5)
-        weekButtons.addArrangedSubview(button6)
-        weekButtons.addArrangedSubview(button7)
-        
-    }
     private func setupWeekInterval(){
         weekIntervalButton.translatesAutoresizingMaskIntoConstraints = false
-        
-        weekInterval.addArrangedSubview(weekIntervalButton)
         let subLabel : UILabel = {
             let label = UILabel()
-            label.text = "주 간격으로 반복"
+            label.text = "매월"
             label.font = UIFont.mpFont16R() // 폰트 크기 설정
             label.textColor = .mpDarkGray
             return label
         }()
+        let subLabel2 : UILabel = {
+            let label = UILabel()
+            label.text = "에 반복"
+            label.font = UIFont.mpFont16R() // 폰트 크기 설정
+            label.textColor = .mpDarkGray
+            return label
+        }()
+       let blank = UIView()
+        NSLayoutConstraint.activate([
+            subLabel.widthAnchor.constraint(equalToConstant: 30 ),
+            subLabel2.widthAnchor.constraint(equalToConstant: 59)
+        ])
+        //weekIBtnConstraint = weekIntervalButton.widthAnchor.constraint(equalToConstant: 38)
+
+        //weekIBtnConstraint!.isActive = true
+        weekIntervalButton.contentEdgeInsets = UIEdgeInsets(top: 3, left: 3, bottom: 3, right: 3)
         weekInterval.addArrangedSubview(subLabel)
-        
+        weekInterval.addArrangedSubview(weekIntervalButton)
+        weekInterval.addArrangedSubview(subLabel2)
+        weekInterval.addArrangedSubview(blank)
+
     }
-    
     
     private func setupRepeatEndDate(){
         let RepeatEndDateLabel : UILabel = {
@@ -204,40 +198,25 @@ class ChooseDayView: UIView {
         repeatEndDate.addArrangedSubview(RepeatEndDateLabel)
         
     }
-    private func setupRepeatEndDateButton(){
-        repeatEndDate2.translatesAutoresizingMaskIntoConstraints = false
-        RepeatEndDate2Label.translatesAutoresizingMaskIntoConstraints = false
-           NSLayoutConstraint.activate([
-            // 높이 설정
-            
-            repeatEndDate2.heightAnchor.constraint(equalToConstant: 38),
-            repeatEndDate2.topAnchor.constraint(equalTo: repeatEndDate.bottomAnchor,constant: 16),
-            repeatEndDate2.leadingAnchor.constraint(equalTo: leadingAnchor),
-            repeatEndDate2.trailingAnchor.constraint(equalTo: trailingAnchor),
-            RepeatEndDate2Label.heightAnchor.constraint(equalToConstant: 38),
-            RepeatEndDate2Label.widthAnchor.constraint(equalToConstant: 28),
-            calChooseButton.widthAnchor.constraint(equalToConstant: 137),
-            calChooseButton.heightAnchor.constraint(equalToConstant: 38),
-            
-           ])
-       
-    }
+    
+    // 반복 종료일 설정
     @objc
     func repeatEndDateButtonTapped(){
         if RepeatEndDateButton.isChecked {
             repeatEndChecked = false
             print("반복 종료일 설정 클릭함 ")
-            delegate?.addEndDay(repeatEndDate2)
-            setupRepeatEndDateButton()
+            //setupRepeatEndDateButton()
+            print("log1")
+            delegate?.addEndDate(repeatEndDate2)
+            print("log2")
+
             setNeedsLayout()
             layoutIfNeeded()
-            
         
         }
         else{
             repeatEndChecked = true
-            repeatEndDate2.removeFromSuperview()
-            delegate?.removeEndDay()
+            delegate?.removeEndDate()
             
         }
     }
@@ -249,8 +228,8 @@ class ChooseDayView: UIView {
             layoutIfNeeded()
             
         }
-                                        
-
-   
+    
+    
+    
 }
